[ADD] map부분 컴포넌트화
